+++ .editorconfig
@@ -1,0 +1,15 @@
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+indent_style = tab
+
+[*.md]
+indent_style = space
+
+[*.{yaml,yml}]
+indent_style = space
+indent_size = 2
