+++ config/default.cjs
@@ -18,6 +18,9 @@ module.exports = {
 	admin: {
 		key: '',
 	},
+	systemApi: {
+		key: '',
+	},
 	geoip: {
 		cache: {
 			ttl: 3 * 24 * 60 * 60 * 1000, // 3 days
